csv2vcard/requirements-gui.txt
PySimpleGUI
ofunctions.threading